refactor(utils): extract SemanticChunkerBase + collapse shared pipeline (#5363)
Consolidates semantic_chunker.py (CPU, 940 lines) and semantic_chunker_gpu.py
(GPU, 572 lines) into a shared base class with hardware-specific embedding
compute + model init as the only divergence.

- NEW: utils/semantic_chunker_base.py with shared sentence splitting (regex
  + abbreviation-aware), cosine-distance computation, percentile-based
  boundary detection, min/max-size-constrained chunk assembly, metadata
  enrichment, fallback chunking, and the public chunk_text() / chunk_document()
  entry points.
- semantic_chunker.py: AutoBotSemanticChunker(SemanticChunkerBase) overrides
  _initialize_model + _compute_embeddings with the adaptive-batch /
  ThreadPoolExecutor CPU path. Retains legacy aliases
  _compute_sentence_embeddings_async / _compute_sentence_embeddings for the
  five sites that call them directly (npu_semantic_search,
  ai_hardware_accelerator, codebase_analytics/*, npu_performance_measurement).
- semantic_chunker_gpu.py: GPUSemanticChunker(SemanticChunkerBase) overrides
  _initialize_model + _compute_embeddings with FP16 + cuDNN + TF32 +
  kernel-warmup + GPU memory pool, and overrides chunk_text to add timing
  telemetry around the shared pipeline.

Behavioral notes:
- GPU pre-refactor used _basic_sentence_split (no abbreviation handling);
  consolidated pipeline uses the abbreviation-aware splitter - a correctness
  improvement for the GPU path.
- GPU pre-refactor _create_chunks_with_boundaries had no min/max size
  handling; consolidated pipeline enforces both size limits on both paths.
- GPU pre-refactor _compute_semantic_distances had a try/except that
  returned default [0.5] distances on failure; in the consolidated pipeline
  exceptions bubble up to chunk_text() and trigger _fallback_chunking
  instead - end result (valid chunks returned) is equivalent.

Public API preserved: AutoBotSemanticChunker, GPUSemanticChunker,
SemanticChunk, get_semantic_chunker(), get_gpu_semantic_chunker(). All
external callers (10 import sites across autobot-backend and
autobot-infrastructure) work unchanged.

Line counts: 940 + 572 = 1512 -> 472 + 331 + 391 = 1194 (saved 318 lines,
~21%). Dedupe is higher in logic terms since the 391-line base replaces
~600 duplicated lines across both originals.

Closes #5363
